Simplify the input states by using const fields
This takes advantage of C++'s const member fields, which need to be
initalized during construction. Once initalized they cannot be modified,
essentially making them "immutable" fields, which we want.

struct's are used as all fields are public and all inheritances are
public.

The only "mutable" field is Inputting's evictedText (formerly
poppedText). Since its setter and getter are trivial and the class is
now a struct, it can just be a field.

This turns most of the state classes into data/value classes and also
shortens the code involved in building them. It's also less error-prone
as their constructors require everything upfront.