Use heuristics for when to ignore FocusOut events
Experiments showed that manually triggered context switch
(activate->reset) actually takes ~200 ms to complete (which is a bit
surprising, by the way). Anything below that latency can only be due to
non-manual triggers like FocusOut. We therefore use heuristics and a
lower threshold (90 ms, or 90,000 microsecnds) to determine when
FocusOut should be ignored if the current state is Inputting and it has
recently evicted (so committed) some overflow text.

Fixes #24.