Fine-tune the handling of A-Z keys
This simplifies the code paths. With the default
(DirectlyOutputUppercase), we also can bypass the handling if the
composing buffer is empty.

Also updates the l10n names since it's the original string that gets
written into the config file, and so we need to put the display strings
in en.po for the en locale.