Prevent sole tone markers from being composed
Previously we allow sole tone markers to be composed on their own. This
created an suboptimal user experience as tone markers might be
inadvertently placed into the composing buffer. Now we require an extra
Space if the intent is really to compose the markers. New test cases are
introduced for this improvement.

Also, independently, this commit fixes a crashing bug: If the composing
buffer starts empty and a non-viable syllable is composed, the state
machine should enter Empty state. Instead, Inputting state was used.
This caused the next Space key to attempt to enter the ChoosingCandidate
state when the builder had nothing to offer, resulting in crash.
Regression tests are added as a result.