Uses tab key to choose next candidate without showing candidate panel.
This feature helps to save time in some cases. For example, when a user
wants to type "…", he or she has to input _ and then use space key
to show the candidate window and then select the second candidae. Using
the tab key feature can let the user just input _ and tab.