Introduce Gramambular2
This new version of Gramambular provides a more streamlined API and
fixes a number of subtle bugs in the past that involve overriding
overlapping nodes. The test suite is also more comprehensive now.

The biggest improvement is in the walk() method. A DAG shortest-path
algorithm is now used, which runs in O(|V| + |E|) time.

In practice, this means walks can be performed with sub-millisecond
latency on a modern machine for normal inputs from an input method
editor.