Fix a bug in how UOM ignores punctuations
Punctuations should be treated as empty words by the UOM, but a bug in
the code missed the check for the frontmost (called the "anterior")
unigram. For example, after typing "！！自會" and choosing "字彙", the
next time the UOM sees "？？自會", it should suggest "字彙", but because
of the bug it didn't happen.

Also addresses the issues found by clang-tidy.