Use values, not references, when passing callbacks
Before this commit, the callbacks that are passed to KeyHandler have
been by-reference. The problem is that, if a callback is owned by a
candidate object, and the key handler invokes the StateCallback with a
committing state, the commit will cause the candidate panel to reset,
destroying the candidate object. At that point the callback reference is
no longer valid, and further use of the callback (still within the key
handler) is therefore no longer sound.

This becomes problematic when a StateCallback is used multiple times and
when the commiting state is involved. To make the code safer, we
therefore update the public functions to take callbacks as values, not
via references.

This doesn't change the private functions of KeyHandler as those are
already referring to the callback object via the value received from the
public functions.