Create a script to derive associated phrases
It uses a new format for the associated phrases, which take a phrase's
reading and value into account. The entries are designed to allow fast
prefix search. Note that the entries are sorted by the entry text's
bytes, not by the score. Actual ranking-by-the-score will be performed
later by the input method.