Retain prefix location for associated phrase
To correctly find where the prefix node is, we need to take into
consideration that (1) the cursor is always *after* the prefix when
Shift-Enter is pressed, but (2) a user may enter the Associated Phrases
state by pressing Shift-Enter within a candidate panel, and the cursor
there does not always match the prefix node location due to the Hanyin
mode requirements.

We therefore make sure that the actual prefix node location is recorded
in the AssociatedPhrase state object, and we now provide an extra
builder method that takes the ChoosingCandidate->AssociatedPhrases state
transition into consideration.