Supports phrases replacement
In fcitx5-mcbopomofo, phrases replacement is enabled if
`phrases-replacement.txt` exists in the user's data path, usually
`$HOME/.local/share/fcitx5/mcbopomofo/phrases-replacement.txt`, and it's
disabled if the file does not exist or is removed.

fcitx5-mcbopomofo does not populate the file with a template; this is
different from the behavior of the macOS version.