Adds a button to open user data in the preferences.
We have already added two menu items in the input menu to help to users
to edit the added user phrases. However, it looks like the input menu is
not available in some desktop environment so users are confused about
where to find the user data. The PR adds an additional button in the
preferences so there will be another entry.