Prompt when reading and composed lengths differ
Before this commit, fcitx5-mcbopomofo only prompts for cursor position
when the composed string's code point count is less than the reading
count. This commit makes sure that it also prompts when the code point
count is bigger than the reading count. This syncs the behavior with
that in the macOS version.