Fix the Shift+[1-9] handling in NumberInput state
The candidate panel were not handling the Shift+[1-9] keys.

This commit also fixes a bug that such Shift keys were not handled when
Shift+Enter was disabled for associated phrases. With the v2 associated
phrases always prompting when such phrases are available, the shift keys
always need to be handled.