Handle candidate keys correctly in the NumberInput
This fixes a regression introduced in https://github.com/openvanilla/fcitx5-mcbopomofo/pull/225
that consumes the Shift+[1-9] keys when there are candidates.
`Key::ascii` alone is not enough to determine that it's produced from a
Shift+[1-9] key, and such keys must be not be consumed in KeyHandler
when the candidate panel is visible. The regression also caused Space
to be consumed unconditionally, casuing Space to fail to work as the
page flip key in the candidate panel.

This also exits the NumberInput state when a non-alphanumeric key
(excluding Esc) is pressed when the number composition buffer is empty.
This fixes the problem that the number prompt ("[數字]") got
force-committed by the client app when fcitx5 was told that the key was
not handled.